add React.js-TypeScript

+++ dojo.sh
@@ -3,7 +3,7 @@ set -e
 rm -rf ~/code/Dojo
 source ~/bin/list_input.sh
 
-templates=( 'Vanilla' 'React.js' 'Docker' 'Node.js')
+templates=( 'Vanilla' 'React.js' 'Docker' 'Node.js' 'React.js-TypeScript')
 list_input "Choose a template:" templates selected_template
  
 if [ "$selected_template" == "Vanilla" ]; then
@@ -36,6 +36,15 @@ if [ "$selected_template" == "Node.js" ]; then
       git clone https://github.com/VKTRN/Source.git ~/code/Dojo &&
       cd ~/code/Dojo &&
       git checkout Node.js
+      rm -rf .gitignore .git
+      npm install &&
+      code .
+fi
+
+if [ "$selected_template" == "React.js-TypeScript" ]; then
+      git clone https://github.com/VKTRN/Source.git ~/code/Dojo &&
+      cd ~/code/Dojo &&
+      git checkout React.js-TypeScript
       rm -rf .gitignore .git
       npm install &&
       code .
